BaseTools: Fix ASL build issue with GCC5 and above
The ASL files are pre-processed using the GCC compiler. Starting with
GCC5, the pre-processor outputs the input file name in lower case. The
build tool 'Trim' expects the input file name in the generated
output to be case compatible with the file system and fails otherwise.
(The symptom being: "Error    6126 -
Input file does not appear to be an ASL or data table source file")

One solution for this problem is to remove the 'Trim --asl-file' line
from the ASL build stage. Removing that 'Trim' does not appear to affect
the build.

Caveat - The GCC5 behaviour mentioned above occurs in a Windows
development environment.  It is not known if the problem might occur on
Linux.

Change-Id: I9206ac8e3cc93084160c0267a4c7946bc3869e3f
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>